% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listbuilder.r
\name{cat.pkg.funs}
\alias{cat.pkg.funs}
\title{Helper function to create whitelists}
\usage{
cat.pkg.funs(pkg = "base")
}
\arguments{
\item{pkg}{the name of the package}
}
\description{
Cats all functions in a package
as a yaml vector
}

\examples{\dontrun{
  cat.pkg.funs("stats")
  cat.pkg.funs("base")
  cat.pkg.funs("dplyr")
  cat.pkg.funs("ggplot2")
}}
