% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist.r
\name{check.blacklist}
\alias{check.blacklist}
\title{A synoym for check.whitelist with different order of the arguments}
\usage{
check.blacklist(call, bl.funs = NULL, bl.vars = NULL, wl.funs = NULL,
  wl.vars = NULL, wl.calls = NULL)
}
\description{
A synoym for check.whitelist with different order of the arguments
}

