% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.r
\name{find.calls}
\alias{find.calls}
\title{Returns a list of calls of the functions listed in fun.names
or variables access}
\usage{
find.calls(call, fun.names = NULL, var.names = NULL, fun.as.var = TRUE,
  return.parent.for.var = TRUE, parent = NULL)
}
\description{
Returns a list of calls of the functions listed in fun.names
or variables access
}

\examples{\dontrun{
  call = quote({
    print(abs(-2*5*log(abs(-1))))
  })
  find.calls(call, fun.names="abs")
  call = parse(text='
    lapply(-5:10,abs)

    lapply(-5:10,abs)
    print(abs(-2*5*log(abs(-1))))
  ')
  find.calls(call, fun.names=c("abs","sum"))
}}
