% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.r
\name{find.funs.except}
\alias{find.funs.except}
\title{Find function calls inside a call object but ignore subcalls listed in ignore.calls}
\usage{
find.funs.except(call, ignore.calls = NULL,
  ignore.names = names(ignore.calls))
}
\arguments{
\item{call}{the call to be analysed}

\item{ignore.calls}{a named list of quoted calls. The names must be the function names of the call. Alternatively, ignore.names can be set}

\item{ignore.names}{just the function names of the ignore.calls}
}
\description{
Find function calls inside a call object but ignore subcalls listed in ignore.calls
}

