% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listbuilder.r
\name{parse.whitelist.yaml}
\alias{parse.whitelist.yaml}
\title{Parse a whitelist yaml file and return it as a character vector}
\usage{
parse.whitelist.yaml(yaml.file = NULL, yaml.text = NULL, yaml.list = NULL)
}
\arguments{
\item{yaml.file}{the file name of the whitelist}

\item{yaml.text}{yaml text as a single character variable}

\item{yaml.list}{an already imported yaml file returned from yaml.load}
}
\value{
a character vector with function names or variable names that are forbidden
}
\description{
For a structure of the yaml file,
see the example whitelists in the lists folder
}

