% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.r
\name{robust.find.funs.except}
\alias{robust.find.funs.except}
\title{Like find.funs.except but also works if call is a (nested) list of calls or an expression}
\usage{
robust.find.funs.except(call, ignore.calls = NULL,
  ignore.names = names(ignore.calls))
}
\description{
Like find.funs.except but also works if call is a (nested) list of calls or an expression
}

