% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.r
\name{robust.find.funs}
\alias{robust.find.funs}
\title{Like find.funs but also works if call is a (nested) list of calls or an expression}
\usage{
robust.find.funs(call)
}
\description{
Like find.funs but also works if call is a (nested) list of calls or an expression
}

