% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.r
\name{robust.find.variables}
\alias{robust.find.variables}
\title{Like find.variables but also works if call is a (nested) list of calls or an expression}
\usage{
robust.find.variables(call)
}
\description{
Like find.variables but also works if call is a (nested) list of calls or an expression
}

