% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelist.r
\name{whitelist.eval}
\alias{whitelist.eval}
\title{Only evaluate expr if it passes check.whitelist}
\usage{
whitelist.eval(expr, envir = parent.frame(), enclos = if (is.list(envir) ||
  is.pairlist(envir)) parent.frame() else baseenv(), wl.funs = NULL,
  wl.vars = NULL, wl.calls = NULL, bl.funs = NULL, bl.vars = NULL)
}
\arguments{
\item{expr}{an R call object that shall be evaluated}

\item{envir}{the environment in which expr shall be evaluated}

\item{enclos}{see help for `eval`}

\item{wl.funs}{a character vector of the function names that are allowed (whitelisted). If NULL ignored.}

\item{wl.vars}{a character vector of the variable names that are allowed (whitelisted). If NULL ignored.}

\item{wl.calls}{a list of explicit calls that are allowed.
For example, one may not generally whitelist the function 'library'
(who knows what can happen if a library has functions with the
same name than some whitelisted function but different behavior)
Yet one may allow the explicit call
`library(dplyr)`. In this case, we could set
 wl.calls = alist(library(dplyr)).
 DOES NOT YET WORK!}

\item{bl.funs}{a character vector of the function names that are forbidden (blacklisted). If NULL ignored.}

\item{bl.vars}{a character vector of the variable names that are forbidden  (blacklisted). If NULL ignored.}
}
\value{
the evaluated expr if it passes check.whitelist, otherwise an error.
}
\description{
if the expression is not cleared by check.whitelist throw an error
}

